Fix worker threads importing prisma (#4839)
