+++ src/lib/addXP.ts
@@ -1,11 +1,12 @@
 import { noOp, Time } from 'e';
+import { convertXPtoLVL, toKMB } from 'oldschooljs/dist/util/util';
 
 import { MAXING_MESSAGE, SupportServer } from '../config';
-import { Events, LEVEL_99_XP, MAX_TOTAL_LEVEL, MAX_XP, skillEmoji } from './constants';
+import { Events, LEVEL_99_XP, MAX_TOTAL_LEVEL, MAX_XP } from './constants';
+import { skillEmoji } from './data/emojis';
 import { AddXpParams } from './minions/types';
 import { prisma } from './settings/prisma';
 import Skills from './skilling/skills';
-import { convertXPtoLVL, toKMB } from './util';
 import { formatOrdinal } from './util/formatOrdinal';
 import { toTitleCase } from './util/toTitleCase';
 import { sendToChannelID } from './util/webhook';
