+++ src/lib/data/buyables/buyables.ts
@@ -1,5 +1,4 @@
 import { Bank } from 'oldschooljs';
-import { Item } from 'oldschooljs/dist/meta/types';
 
 import { soteSkillRequirements } from '../../../mahoji/lib/abstracted_commands/zalcanoCommand';
 import { chompyHats, MAX_QP } from '../../constants';
@@ -7,8 +6,8 @@ import { CombatCannonItemBank } from '../../minions/data/combatConstants';
 import { Favours } from '../../minions/data/kourendFavour';
 import { MinigameName } from '../../settings/settings';
 import { Skills } from '../../types';
-import getOSItem from '../../util/getOSItem';
 import itemID from '../../util/itemID';
+import { allTeamCapes } from '../misc';
 import { aerialFishBuyables } from './aerialFishBuyables';
 import { canifisClothes } from './canifisClothes';
 import { capeBuyables } from './capes';
@@ -1050,10 +1049,6 @@ for (const [chompyHat, qty] of chompyHats) {
 	});
 }
 
-export const allTeamCapes: Item[] = [];
-for (let i = 1; i < 51; i++) {
-	allTeamCapes.push(getOSItem(`Team-${i} cape`));
-}
 for (const cape of allTeamCapes) {
 	Buyables.push({
 		name: cape.name,
