+++ src/lib/data/emojis.ts
@@ -1,0 +1,30 @@
+export const skillEmoji = {
+	runecraft: '<:runecraft:630911040435257364>',
+	firemaking: '<:firemaking:630911040175210518>',
+	thieving: '<:thieving:630910829352452123>',
+	mining: '<:mining:630911040128811010>',
+	ranged: '<:ranged:630911040258834473>',
+	construction: '<:construction:630911040493715476>',
+	smithing: '<:smithing:630911040452034590>',
+	herblore: '<:herblore:630911040535658496>',
+	attack: '<:attack:630911039969427467>',
+	strength: '<:strength:630911040481263617>',
+	defence: '<:defence:630911040393052180>',
+	fishing: '<:fishing:630911040091193356>',
+	hitpoints: '<:hitpoints:630911040460292108>',
+	total: '<:xp:630911040510623745>',
+	overall: '<:xp:630911040510623745>',
+	magic: '<:magic:630911040334331917>',
+	crafting: '<:crafting:630911040460161047>',
+	agility: '<:agility:630911040355565568>',
+	fletching: '<:fletching:630911040544309258>',
+	cooking: '<:cooking:630911040426868756>',
+	farming: '<:farming:630911040355565599>',
+	slayer: '<:slayer:630911040560824330>',
+	prayer: '<:prayer:630911040426868746>',
+	woodcutting: '<:woodcutting:630911040099450892>',
+	hunter: '<:hunter:630911040166559784>',
+	cml: '<:CrystalMathLabs:364657225249062912>',
+	clock: '<:ehpclock:352323705210142721>',
+	combat: '<:combat:802136963956080650>'
+};
