+++ src/lib/minions/data/sepulchre.ts
@@ -1,9 +1,8 @@
-import { randInt, Time } from 'e';
+import { randInt, roll, Time } from 'e';
 import { Bank } from 'oldschooljs';
 import LootTable from 'oldschooljs/dist/structures/LootTable';
 
 import { ItemBank } from '../../types';
-import { roll } from '../../util';
 import resolveItems from '../../util/resolveItems';
 
 const LowTierCoffin = new LootTable()
