+++ src/lib/patreon.ts
@@ -2,12 +2,11 @@ import { Time } from 'e';
 import fetch from 'node-fetch';
 
 import { patreonConfig, production } from '../config';
-import { mahojiUserSettingsUpdate } from '../mahoji/settingsUpdate';
 import { cacheBadges } from './badges';
 import { BadgesEnum, BitField, Channel, PatronTierID, PerkTier } from './constants';
 import { fetchSponsors, getUserIdFromGithubID } from './http/util';
 import backgroundImages from './minions/data/bankBackgrounds';
-import { getUsersPerkTier } from './MUser';
+import { getUsersPerkTier, mahojiUserSettingsUpdate } from './MUser';
 import { roboChimpUserFetch } from './roboChimp';
 import { Patron } from './types';
 import { logError } from './util/logError';
