+++ src/lib/util.ts
@@ -1,4 +1,3 @@
-import { bold } from '@discordjs/builders';
 import { Stopwatch } from '@sapphire/stopwatch';
 import {
 	BaseMessageOptions,
@@ -6,7 +5,6 @@ import {
 	ButtonInteraction,
 	CacheType,
 	Channel,
-	ChannelType,
 	Collection,
 	CollectorFilter,
 	ComponentType,
@@ -24,35 +22,31 @@ import {
 	time,
 	User as DJSUser
 } from 'discord.js';
-import { calcWhatPercent, chunk, isObject, objectEntries, Time } from 'e';
+import { chunk, isObject, objectEntries, Time } from 'e';
 import { CommandResponse } from 'mahoji/dist/lib/structures/ICommand';
 import murmurHash from 'murmurhash';
 import { gzip } from 'node:zlib';
 import { Bank } from 'oldschooljs';
 import { ItemBank } from 'oldschooljs/dist/meta/types';
-import Items from 'oldschooljs/dist/structures/Items';
-import { bool, integer, MersenneTwister19937, nodeCrypto, real, shuffle } from 'random-js';
+import { bool, integer, nodeCrypto, real } from 'random-js';
 
-import { ADMIN_IDS, OWNER_IDS, production, SupportServer } from '../config';
-import { badgesCache, BitField, skillEmoji, usernameCache } from './constants';
+import { ADMIN_IDS, OWNER_IDS, SupportServer } from '../config';
+import { badgesCache, BitField, usernameCache } from './constants';
 import { DefenceGearStat, GearSetupType, GearSetupTypes, GearStat, OffenceGearStat } from './gear/types';
-import { Consumable } from './minions/types';
+import type { Consumable } from './minions/types';
 import { MUserClass } from './MUser';
 import { PaginatedMessage } from './PaginatedMessage';
-import { POHBoosts } from './poh';
+import type { POHBoosts } from './poh';
 import { SkillsEnum } from './skilling/types';
-import { ArrayItemsResolved, Skills } from './types';
-import {
+import type { Skills } from './types';
+import type {
 	GroupMonsterActivityTaskOptions,
 	NexTaskOptions,
 	RaidsOptions,
 	TheatreOfBloodTaskOptions
 } from './types/minions';
-import { CACHED_ACTIVE_USER_IDS } from './util/cachedUserIDs';
 import { getItem } from './util/getOSItem';
 import itemID from './util/itemID';
-import { logError } from './util/logError';
-import { toTitleCase } from './util/toTitleCase';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const emojiRegex = require('emoji-regex');
@@ -107,27 +101,6 @@ export function formatItemStackQuantity(quantity: number) {
 	return quantity.toString();
 }
 
-export function formatDuration(ms: number, short = false) {
-	if (ms < 0) ms = -ms;
-	const time = {
-		day: Math.floor(ms / 86_400_000),
-		hour: Math.floor(ms / 3_600_000) % 24,
-		minute: Math.floor(ms / 60_000) % 60,
-		second: Math.floor(ms / 1000) % 60
-	};
-	const shortTime = {
-		d: Math.floor(ms / 86_400_000),
-		h: Math.floor(ms / 3_600_000) % 24,
-		m: Math.floor(ms / 60_000) % 60,
-		s: Math.floor(ms / 1000) % 60
-	};
-	let nums = Object.entries(short ? shortTime : time).filter(val => val[1] !== 0);
-	if (nums.length === 0) return '1 second';
-	return nums
-		.map(([key, val]) => `${val}${short ? '' : ' '}${key}${val === 1 || short ? '' : 's'}`)
-		.join(short ? '' : ', ');
-}
-
 export function isWeekend() {
 	const currentDate = new Date(Date.now() - Time.Hour * 6);
 	return [6, 0].includes(currentDate.getDay());
@@ -163,10 +136,6 @@ export function roll(max: number) {
 	return rand(1, max) === 1;
 }
 
-export function itemNameFromID(itemID: number | string) {
-	return Items.get(itemID)?.name;
-}
-
 const rawEmojiRegex = emojiRegex();
 
 export function stripEmojis(str: string) {
@@ -182,18 +151,6 @@ export const anglerBoosts = [
 
 export function isValidGearSetup(str: string): str is GearSetupType {
 	return GearSetupTypes.includes(str as any);
-}
-
-/**
- * Adds random variation to a number. For example, if you pass 10%, it can at most lower the value by 10%,
- * or increase it by 10%, and everything in between.
- * @param value The value to add variation too.
- * @param percentage The max percentage to fluctuate the value by, in both negative/positive.
- */
-export function randomVariation(value: number, percentage: number) {
-	const lowerLimit = value * (1 - percentage / 100);
-	const upperLimit = value * (1 + percentage / 100);
-	return randFloat(lowerLimit, upperLimit);
 }
 
 export function isGroupActivity(data: any): data is GroupMonsterActivityTaskOptions {
@@ -217,16 +174,6 @@ export function getSupportGuild(): Guild | null {
 	const guild = globalClient.guilds.cache.get(SupportServer);
 	if (!guild) return null;
 	return guild;
-}
-
-export function normal(mu = 0, sigma = 1, nsamples = 6) {
-	let run_total = 0;
-
-	for (let i = 0; i < nsamples; i++) {
-		run_total += Math.random();
-	}
-
-	return (sigma * (run_total - nsamples / 2)) / (nsamples / 2) + mu;
 }
 
 /**
@@ -272,18 +219,6 @@ export function skillsMeetRequirements(skills: Skills, requirements: Skills) {
 		}
 	}
 	return true;
-}
-
-export function formatItemReqs(items: ArrayItemsResolved) {
-	const str = [];
-	for (const item of items) {
-		if (Array.isArray(item)) {
-			str.push(item.map(itemNameFromID).join(' OR '));
-		} else {
-			str.push(itemNameFromID(item));
-		}
-	}
-	return str.join(', ');
 }
 
 export function formatItemCosts(consumable: Consumable, timeToFinish: number) {
@@ -337,34 +272,6 @@ export function formatMissingItems(consumables: Consumable[], timeToFinish: numb
 	return str.join(', ');
 }
 
-export function formatSkillRequirements(reqs: Record<string, number>, emojis = true) {
-	let arr = [];
-	for (const [name, num] of objectEntries(reqs)) {
-		arr.push(`${emojis ? ` ${(skillEmoji as any)[name]} ` : ''}**${num}** ${toTitleCase(name)}`);
-	}
-	return arr.join(', ');
-}
-
-export function formatItemBoosts(items: ItemBank[]) {
-	const str = [];
-	for (const itemSet of items) {
-		const itemEntries = Object.entries(itemSet);
-		const multiple = itemEntries.length > 1;
-		const bonusStr = [];
-
-		for (const [itemID, boostAmount] of itemEntries) {
-			bonusStr.push(`${boostAmount}% for ${itemNameFromID(parseInt(itemID))}`);
-		}
-
-		if (multiple) {
-			str.push(`(${bonusStr.join(' OR ')})`);
-		} else {
-			str.push(bonusStr.join(''));
-		}
-	}
-	return str.join(', ');
-}
-
 export function formatPohBoosts(boosts: POHBoosts) {
 	const bonusStr = [];
 	const slotStr = [];
@@ -409,38 +316,6 @@ export type PaginatedMessagePage = MessageEditOptions;
 export async function makePaginatedMessage(channel: TextChannel, pages: PaginatedMessagePage[], target?: string) {
 	const m = new PaginatedMessage({ pages, channel });
 	return m.run(target ? [target] : undefined);
-}
-
-export function assert(condition: boolean, desc?: string, context?: Record<string, string>) {
-	if (!condition) {
-		if (production) {
-			logError(new Error(desc ?? 'Failed assertion'), context);
-		} else {
-			throw new Error(desc ?? 'Failed assertion');
-		}
-	}
-}
-
-export function calcDropRatesFromBank(bank: Bank, iterations: number, uniques: number[]) {
-	let result = [];
-	let uniquesReceived = 0;
-	for (const [item, qty] of bank.items().sort((a, b) => a[1] - b[1])) {
-		if (uniques.includes(item.id)) {
-			uniquesReceived += qty;
-		}
-		const rate = Math.round(iterations / qty);
-		if (rate < 2) continue;
-		let { name } = item;
-		if (uniques.includes(item.id)) name = bold(name);
-		result.push(`${qty}x ${name} (1 in ${rate})`);
-	}
-	result.push(
-		`\n**${uniquesReceived}x Uniques (1 in ${Math.round(iterations / uniquesReceived)} which is ${calcWhatPercent(
-			uniquesReceived,
-			iterations
-		)}%)**`
-	);
-	return result.join(', ');
 }
 
 export function convertPercentChance(percent: number) {
@@ -515,29 +390,6 @@ export function removeMarkdownEmojis(str: string) {
 }
 export { cleanString, stringMatches } from './util/cleanString';
 
-export function clamp(val: number, min: number, max: number) {
-	return Math.min(max, Math.max(min, val));
-}
-
-export function calcPerHour(value: number, duration: number) {
-	return (value / (duration / Time.Minute)) * 60;
-}
-
-export function removeFromArr<T>(arr: T[] | readonly T[], item: T) {
-	return arr.filter(i => i !== item);
-}
-
-/**
- * Scale percentage exponentially
- *
- * @param decay Between 0.01 and 0.05; bigger means more penalty.
- * @param percent The percent to scale
- * @returns percent
- */
-export function exponentialPercentScale(percent: number, decay = 0.021) {
-	return 100 * Math.pow(Math.E, -decay * (100 - percent));
-}
-
 export function discrimName(user: DJSUser) {
 	return `${user.username}#${user.discriminator}`;
 }
@@ -605,11 +457,6 @@ export function getUsername(id: string | bigint, withBadges: boolean = true) {
 	return username;
 }
 
-export function shuffleRandom<T>(input: number, arr: readonly T[]): T[] {
-	const engine = MersenneTwister19937.seed(input);
-	return shuffle(engine, [...arr]);
-}
-
 export function makeComponents(components: ButtonBuilder[]): InteractionReplyOptions['components'] {
 	return chunk(components, 5).map(i => ({ components: i, type: ComponentType.ActionRow }));
 }
@@ -630,13 +477,6 @@ export function validateItemBankAndThrow(input: any): input is ItemBank {
 	return true;
 }
 
-export function hasSkillReqs(user: MUser, reqs: Skills): [boolean, string | null] {
-	const hasReqs = user.hasSkillReqs(reqs);
-	if (!hasReqs) {
-		return [false, formatSkillRequirements(reqs)];
-	}
-	return [true, null];
-}
 type test = CollectorFilter<
 	[
 		ButtonInteraction<CacheType> | SelectMenuInteraction<CacheType>,
@@ -675,110 +515,6 @@ export async function runTimedLoggedFn(name: string, fn: () => Promise<unknown>)
 	debugLog(`Finished ${name} in ${stopwatch.toString()}`);
 }
 
-const emojiServers = new Set([
-	'342983479501389826',
-	'940758552425955348',
-	'869497440947015730',
-	'324127314361319427',
-	'363252822369894400',
-	'395236850119213067',
-	'325950337271857152',
-	'395236894096621568'
-]);
-
-export function memoryAnalysis() {
-	let guilds = globalClient.guilds.cache.size;
-	let emojis = 0;
-	let channels = globalClient.channels.cache.size;
-	let voiceChannels = 0;
-	let guildTextChannels = 0;
-	let roles = 0;
-	for (const guild of globalClient.guilds.cache.values()) {
-		emojis += guild.emojis.cache.size;
-		for (const channel of guild.channels.cache.values()) {
-			if (channel.type === ChannelType.GuildVoice) voiceChannels++;
-			if (channel.type === ChannelType.GuildText) guildTextChannels++;
-		}
-		roles += guild.roles.cache.size;
-	}
-	return {
-		guilds,
-		emojis,
-		channels,
-		voiceChannels,
-		guildTextChannels,
-		roles,
-		activeIDs: CACHED_ACTIVE_USER_IDS.size
-	};
-}
-
-export function cacheCleanup() {
-	debugLog('Cache Cleanup', {
-		type: 'CACHE_CLEANUP'
-	});
-	return runTimedLoggedFn('Cache Cleanup', async () => {
-		await runTimedLoggedFn('Clear Channels', async () => {
-			for (const channel of globalClient.channels.cache.values()) {
-				if (channel.type === ChannelType.GuildVoice || channel.type === ChannelType.GuildCategory) {
-					globalClient.channels.cache.delete(channel.id);
-				}
-				if (channel.type === ChannelType.GuildText) {
-					channel.threads.cache.clear();
-					// @ts-ignore ignore
-					delete channel.topic;
-					// @ts-ignore ignore
-					delete channel.rateLimitPerUser;
-					// @ts-ignore ignore
-					delete channel.nsfw;
-					// @ts-ignore ignore
-					delete channel.parentId;
-					// @ts-ignore ignore
-					delete channel.name;
-					// @ts-ignore ignore
-					channel.lastMessageId = null;
-					// @ts-ignore ignore
-					channel.lastPinTimestamp = null;
-				}
-			}
-		});
-
-		await runTimedLoggedFn('Guild Emoji/Roles/Member cache clear', async () => {
-			for (const guild of globalClient.guilds.cache.values()) {
-				if (emojiServers.has(guild.id)) continue;
-				guild.emojis.cache.clear();
-				for (const member of guild.members.cache.values()) {
-					if (!CACHED_ACTIVE_USER_IDS.has(member.user.id)) {
-						guild.members.cache.delete(member.user.id);
-					}
-				}
-				for (const channel of guild.channels.cache.values()) {
-					if (channel.type === ChannelType.GuildVoice || channel.type === ChannelType.GuildNewsThread) {
-						guild.channels.cache.delete(channel.id);
-					}
-				}
-				for (const role of guild.roles.cache.values()) {
-					// @ts-ignore ignore
-					delete role.managed;
-					// @ts-ignore ignore
-					delete role.name;
-					// @ts-ignore ignore
-					delete role.tags;
-					// @ts-ignore ignore
-					delete role.icon;
-					// @ts-ignore ignore
-					delete role.unicodeEmoji;
-					// @ts-ignore ignore
-					delete role.rawPosition;
-					// @ts-ignore ignore
-					delete role.color;
-					// @ts-ignore ignore
-					delete role.hoist;
-				}
-			}
-		});
-	});
-}
-
 export function isFunction(input: unknown): input is Function {
 	return typeof input === 'function';
 }
@@ -814,3 +550,6 @@ export function getInteractionTypeName(type: InteractionType) {
 export function isModOrAdmin(user: MUser) {
 	return [...OWNER_IDS, ...ADMIN_IDS].includes(user.id) || user.bitfield.includes(BitField.isModerator);
 }
+
+export { assert } from './util/logError';
+export * from './util/smallUtils';
