+++ src/lib/util/resolveItems.ts
@@ -1,6 +1,6 @@
 import { Items } from 'oldschooljs';
 
-import { ArrayItemsResolvable, ArrayItemsResolved } from '../types';
+import type { ArrayItemsResolvable, ArrayItemsResolved } from '../types';
 
 // Resolve an array of item IDs or names into an array of item IDs
 export default function resolveItems(_itemArray: string | number | (string | number)[]): number[] {
