+++ src/lib/util/smallUtils.ts
@@ -1,0 +1,119 @@
+import { objectEntries, Time } from 'e';
+import { Items } from 'oldschooljs';
+import { ItemBank } from 'oldschooljs/dist/meta/types';
+import { MersenneTwister19937, shuffle } from 'random-js';
+
+import { skillEmoji } from '../data/emojis';
+import type { ArrayItemsResolved, Skills } from '../types';
+import { toTitleCase } from './toTitleCase';
+
+export function itemNameFromID(itemID: number | string) {
+	return Items.get(itemID)?.name;
+}
+
+export function formatItemReqs(items: ArrayItemsResolved) {
+	const str = [];
+	for (const item of items) {
+		if (Array.isArray(item)) {
+			str.push(item.map(itemNameFromID).join(' OR '));
+		} else {
+			str.push(itemNameFromID(item));
+		}
+	}
+	return str.join(', ');
+}
+
+export function formatItemBoosts(items: ItemBank[]) {
+	const str = [];
+	for (const itemSet of items) {
+		const itemEntries = Object.entries(itemSet);
+		const multiple = itemEntries.length > 1;
+		const bonusStr = [];
+
+		for (const [itemID, boostAmount] of itemEntries) {
+			bonusStr.push(`${boostAmount}% for ${itemNameFromID(parseInt(itemID))}`);
+		}
+
+		if (multiple) {
+			str.push(`(${bonusStr.join(' OR ')})`);
+		} else {
+			str.push(bonusStr.join(''));
+		}
+	}
+	return str.join(', ');
+}
+
+export function clamp(val: number, min: number, max: number) {
+	return Math.min(max, Math.max(min, val));
+}
+
+export function calcPerHour(value: number, duration: number) {
+	return (value / (duration / Time.Minute)) * 60;
+}
+
+export function removeFromArr<T>(arr: T[] | readonly T[], item: T) {
+	return arr.filter(i => i !== item);
+}
+
+export function formatDuration(ms: number, short = false) {
+	if (ms < 0) ms = -ms;
+	const time = {
+		day: Math.floor(ms / 86_400_000),
+		hour: Math.floor(ms / 3_600_000) % 24,
+		minute: Math.floor(ms / 60_000) % 60,
+		second: Math.floor(ms / 1000) % 60
+	};
+	const shortTime = {
+		d: Math.floor(ms / 86_400_000),
+		h: Math.floor(ms / 3_600_000) % 24,
+		m: Math.floor(ms / 60_000) % 60,
+		s: Math.floor(ms / 1000) % 60
+	};
+	let nums = Object.entries(short ? shortTime : time).filter(val => val[1] !== 0);
+	if (nums.length === 0) return '1 second';
+	return nums
+		.map(([key, val]) => `${val}${short ? '' : ' '}${key}${val === 1 || short ? '' : 's'}`)
+		.join(short ? '' : ', ');
+}
+
+export function formatSkillRequirements(reqs: Record<string, number>, emojis = true) {
+	let arr = [];
+	for (const [name, num] of objectEntries(reqs)) {
+		arr.push(`${emojis ? ` ${(skillEmoji as any)[name]} ` : ''}**${num}** ${toTitleCase(name)}`);
+	}
+	return arr.join(', ');
+}
+
+export function hasSkillReqs(user: MUser, reqs: Skills): [boolean, string | null] {
+	const hasReqs = user.hasSkillReqs(reqs);
+	if (!hasReqs) {
+		return [false, formatSkillRequirements(reqs)];
+	}
+	return [true, null];
+}
+
+/**
+ * Scale percentage exponentially
+ *
+ * @param decay Between 0.01 and 0.05; bigger means more penalty.
+ * @param percent The percent to scale
+ * @returns percent
+ */
+export function exponentialPercentScale(percent: number, decay = 0.021) {
+	return 100 * Math.pow(Math.E, -decay * (100 - percent));
+}
+
+export function normal(mu = 0, sigma = 1, nsamples = 6) {
+	let run_total = 0;
+
+	for (let i = 0; i < nsamples; i++) {
+		run_total += Math.random();
+	}
+
+	return (sigma * (run_total - nsamples / 2)) / (nsamples / 2) + mu;
+}
+
+export function shuffleRandom<T>(input: number, arr: readonly T[]): T[] {
+	const engine = MersenneTwister19937.seed(input);
+	return shuffle(engine, [...arr]);
+}
