+++ src/lib/util/statsEmbed.ts
@@ -2,7 +2,7 @@ import { Embed } from '@discordjs/builders';
 import { Player } from 'oldschooljs';
 import { CluesScore, SkillScore, SkillsScore } from 'oldschooljs/dist/meta/types';
 
-import { skillEmoji } from '../constants';
+import { skillEmoji } from '../data/emojis';
 import { toTitleCase } from './toTitleCase';
 
 export function statsEmbed({
