+++ src/lib/workers/kill.worker.ts
@@ -3,12 +3,12 @@ import '../data/itemAliases';
 import { Bank, Misc, Monsters } from 'oldschooljs';
 
 import { handleNexKills } from '../simulation/nex';
-import { calcDropRatesFromBank } from '../util';
+import { calcDropRatesFromBank } from '../util/calcDropRatesFromBank';
 import { stringMatches } from '../util/cleanString';
 import resolveItems from '../util/resolveItems';
-import { KillWorkerArgs, KillWorkerReturn } from '.';
+import type { KillWorkerArgs, KillWorkerReturn } from '.';
 
-export default ({ quantity, bossName, limit, catacombs, onTask }: KillWorkerArgs): KillWorkerReturn => {
+export default async ({ quantity, bossName, limit, catacombs, onTask }: KillWorkerArgs): KillWorkerReturn => {
 	const osjsMonster = Monsters.find(mon => mon.aliases.some(alias => stringMatches(alias, bossName)));
 
 	if (osjsMonster) {
@@ -38,6 +38,7 @@ export default ({ quantity, bossName, limit, catacombs, onTask }: KillWorkerArgs
 		if (quantity > 3000) {
 			return { error: 'I can only kill a maximum of 3k Nex a time!' };
 		}
+
 		const loot = handleNexKills({
 			quantity,
 			team: [
