+++ src/mahoji/lib/events.ts
@@ -8,7 +8,8 @@ import { DISABLED_COMMANDS } from '../../lib/constants';
 import { initCrons } from '../../lib/crons';
 import { prisma } from '../../lib/settings/prisma';
 import { initTickers } from '../../lib/tickers';
-import { cacheCleanup, runTimedLoggedFn } from '../../lib/util';
+import { runTimedLoggedFn } from '../../lib/util';
+import { cacheCleanup } from '../../lib/util/cachedUserIDs';
 import { mahojiClientSettingsFetch } from '../../lib/util/clientSettings';
 import { syncLinkedAccounts } from '../../lib/util/linkedAccountsUtil';
 import { cacheUsernames } from '../commands/leaderboard';
