+++ src/mahoji/mahojiSettings.ts
@@ -18,6 +18,7 @@ import { CLIENT_ID } from '../config';
 import { SILENT_ERROR } from '../lib/constants';
 import { evalMathExpression } from '../lib/expressionParser';
 import { KillableMonster } from '../lib/minions/types';
+import { mahojiUserSettingsUpdate } from '../lib/MUser';
 import { prisma } from '../lib/settings/prisma';
 import { Rune } from '../lib/skilling/skills/runecraft';
 import { hasGracefulEquipped, readableStatName } from '../lib/structures/Gear';
@@ -32,7 +33,6 @@ import {
 } from '../lib/util';
 import { deferInteraction, interactionReply } from '../lib/util/interactionReply';
 import resolveItems from '../lib/util/resolveItems';
-import { mahojiUserSettingsUpdate } from './settingsUpdate';
 
 export function mahojiParseNumber({
 	input,
